Creating more entities

+++ src/main/java/railway/booking/app/repository/CoachRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import railway.booking.app.entities.Coach;
 
-public interface CoachRepository extends JpaRepository<Coach, Integer> {
+public interface CoachRepository extends JpaRepository<Coach, Long> {
 
 }
